LayoutTests/http/tests/websocket/tests/hixie76/undefined-attributes.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Test WebSocket attributes which should not be defined when hixie-76 protocol is used.");

if (window.testRunner)
    testRunner.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 1);

var ws = new WebSocket("ws://127.0.0.1:8880/websocket/tests/hixie76/simple");

shouldBeUndefined("ws.binaryType");
shouldBeUndefined("ws.extensions");
shouldBeUndefined("ws.protocol");

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
